requirements.txt
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.4
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.1.1
reportlab==4.4.10
openai
requests==2.32.4
urllib3==2.5.0
Werkzeug==3.1.3
